requirements.txt
absl-py @ file:///C:/ci/absl-py_1603893176799/work
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
asgiref==3.2.3
astor==0.8.1
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.11.8
cffi @ file:///C:/ci/cffi_1606255207413/work
chardet @ file:///C:/ci/chardet_1605303259695/work
click==7.1.2
colorama==0.4.4
cryptography @ file:///C:/ci/cryptography_1605544557248/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.2
django-crispy-forms==1.8.1
django-restframework==0.0.1
entrypoints==0.3
face-recognition-models==0.3.0
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1600960338579/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab==2.2.6
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown @ file:///C:/ci/markdown_1605111187600/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.3.4
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
nltk==3.5
notebook @ file:///C:/ci/notebook_1601494955742/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1606930849755/work
pandas==1.2.2
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit==3.0.16
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.8.0
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz==2019.3
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
regex==2020.11.13
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-learn==0.24.1
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.2
sqlparse==0.3.0
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
touch==2020.12.3
tqdm==4.56.0
traitlets==5.0.5
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
values==2020.12.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
